Update camerafollow.cs
updated so that the x, y and z axis of the camera follow action are relative to the GameObject. Which in Unity is set to self. This way when people using the tutorial but started in a different x,y,z canvas dont get camera trouble.

Enjoying your MVA presentation :-))